[Feat] - Interview 조회수 count 기능 구현 (#131)
* refactor: MemberAuth 파라미터를 받을 때 필수 여부를 어노테이션으로 지정하도록 변경

* chore : X-Forwarded-For 헤더에서 첫번쨰로 신뢰하지 않는 IP를 X-Real-IP 로 설정

* refactor: like_count, view_count 를 Long으로 변경

* feat: 클라이언트의 IP를 주입해 주는 ArgumentResolver 추가

* feat: 인터뷰의 조회수를 count 하는 기능 구현

* fix: like_count, view_count 타입 변경하는 flyway 스크립트 추가

* test: docs 프로필 테스트에서 create-drop 및 flyway를 사용하지 않도록 변경

* refactor: expire 실패 시에만 SETNX 를 하도록 변경

* test: 인터뷰 조회수 동시성 테스트 추가

* refactor: 인증되지 않은 MemberAuth 정적 캐싱(일종의 Null Object Pattern 적용)

* refactor: 레디스 키 네이밍 변경

* refactor: X-Real-IP 헤더가 비어 있는 경우 0.0.0.0 으로 처리

+++ src/main/java/com/samhap/kokomen/interview/domain/Interview.java
@@ -58,10 +58,10 @@ public class Interview extends BaseEntity {
     private Integer totalScore;
 
     @Column(name = "like_count", nullable = false)
-    private Integer likeCount;
+    private Long likeCount;
 
     @Column(name = "view_count", nullable = false)
-    private Integer viewCount;
+    private Long viewCount;
 
     public Interview(
             Long id,
@@ -71,8 +71,8 @@ public class Interview extends BaseEntity {
             InterviewState interviewState,
             String totalFeedback,
             Integer totalScore,
-            Integer likeCount,
-            Integer viewCount
+            Long likeCount,
+            Long viewCount
     ) {
         validateMaxQuestionCount(maxQuestionCount);
         this.id = id;
@@ -87,7 +87,7 @@ public class Interview extends BaseEntity {
     }
 
     public Interview(Member member, RootQuestion rootQuestion, Integer maxQuestionCount) {
-        this(null, member, rootQuestion, maxQuestionCount, InterviewState.IN_PROGRESS, null, null, 0, 0);
+        this(null, member, rootQuestion, maxQuestionCount, InterviewState.IN_PROGRESS, null, null, 0L, 0L);
     }
 
     private void validateMaxQuestionCount(Integer maxQuestionCount) {
